requirements.txt
certifi==2022.6.15
polygon-api-client==1.0.2
urllib3==1.26.9
websockets==10.3
certifi==2022.6.15
polygon-api-client==1.0.2
python-dotenv==0.20.0
urllib3==1.26.9
websockets==10.3
